Fix various errors/bugs in PostgreSQL driver, speedup SQL queries by omitting write/change to fields that are not changing their value
MFM: 0 days
Bug-ID: none
Bug-Reported-By: none
Security: none

+++ src/pgsql_drv.c
@@ -1,4 +1,4 @@
-/* $Id: pgsql_drv.c,v 1.738 2010/04/30 21:02:31 sbajic Exp $ */
+/* $Id: pgsql_drv.c,v 1.739 2010/05/07 21:14:36 sbajic Exp $ */
 
 /*
  DSPAM
@@ -221,7 +221,7 @@ _pgsql_drv_get_spamtotals (DSPAM_CTX * CTX)
   if (gid != uid)
     snprintf (query, sizeof (query),
 	      "SELECT uid,spam_learned,innocent_learned,"
-	      "spam_misclassified, innocent_misclassified,"
+	      "spam_misclassified,innocent_misclassified,"
 	      "spam_corpusfed,innocent_corpusfed,"
 	      "spam_classified,innocent_classified"
 	      " FROM dspam_stats WHERE uid IN ('%d','%d')",
@@ -398,6 +398,7 @@ _pgsql_drv_set_spamtotals (DSPAM_CTX * CTX)
   char *name;
   char query[1024];
   PGresult *result;
+  int update_any = 0;
   struct _ds_spam_totals user;
   result = NULL;
 
@@ -467,62 +468,130 @@ _pgsql_drv_set_spamtotals (DSPAM_CTX * CTX)
     result = PQexec(s->dbh, query);
   }
 
-  if ( s->control_totals.innocent_learned != 0 || (PQresultStatus(result) != PGRES_COMMAND_OK && PQresultStatus(result) != PGRES_NONFATAL_ERROR) )
+  if ( s->control_totals.innocent_learned != 0 || result == NULL || (PQresultStatus(result) != PGRES_COMMAND_OK && PQresultStatus(result) != PGRES_NONFATAL_ERROR) )
   {
     if (result) PQclear(result);
 
-    snprintf (query, sizeof (query),
-              "UPDATE dspam_stats SET spam_learned=spam_learned%s%d,"
-              "innocent_learned=innocent_learned%s%d,"
-              "spam_misclassified=spam_misclassified%s%d,"
-              "innocent_misclassified=innocent_misclassified%s%d,"
-              "spam_corpusfed=spam_corpusfed%s%d,"
-              "innocent_corpusfed=innocent_corpusfed%s%d,"
-              "spam_classified=spam_classified%s%d,"
-              "innocent_classified=innocent_classified%s%d"
-              " WHERE uid=%d",
-              (CTX->totals.spam_learned >
-               s->control_totals.spam_learned) ? "+" : "-",
-              abs (CTX->totals.spam_learned -
-                   s->control_totals.spam_learned),
-              (CTX->totals.innocent_learned >
-               s->control_totals.innocent_learned) ? "+" : "-",
-              abs (CTX->totals.innocent_learned -
-                   s->control_totals.innocent_learned),
-              (CTX->totals.spam_misclassified >
-               s->control_totals.spam_misclassified) ? "+" : "-",
-              abs (CTX->totals.spam_misclassified -
-                   s->control_totals.spam_misclassified),
-              (CTX->totals.innocent_misclassified >
-               s->control_totals.innocent_misclassified) ? "+" : "-",
-              abs (CTX->totals.innocent_misclassified -
-                   s->control_totals.innocent_misclassified),
-              (CTX->totals.spam_corpusfed >
-               s->control_totals.spam_corpusfed) ? "+" : "-",
-              abs (CTX->totals.spam_corpusfed -
-                   s->control_totals.spam_corpusfed),
-              (CTX->totals.innocent_corpusfed >
-               s->control_totals.innocent_corpusfed) ? "+" : "-",
-              abs (CTX->totals.innocent_corpusfed -
-                   s->control_totals.innocent_corpusfed),
-              (CTX->totals.spam_classified >
-               s->control_totals.spam_classified) ? "+" : "-",
-              abs (CTX->totals.spam_classified -
-                  s->control_totals.spam_classified),
-              (CTX->totals.innocent_classified >
-               s->control_totals.innocent_classified) ? "+" : "-",
-              abs (CTX->totals.innocent_classified -
-                  s->control_totals.innocent_classified), (int) p->pw_uid);
+    /* Do not update stats if all values are zero (aka: no update needed) */
+    if (!(abs(CTX->totals.spam_learned           - s->control_totals.spam_learned) == 0 &&
+          abs(CTX->totals.innocent_learned       - s->control_totals.innocent_learned) == 0 &&
+          abs(CTX->totals.spam_misclassified     - s->control_totals.spam_misclassified) == 0 &&
+          abs(CTX->totals.innocent_misclassified - s->control_totals.innocent_misclassified) == 0 &&
+          abs(CTX->totals.spam_corpusfed         - s->control_totals.spam_corpusfed) == 0 &&
+          abs(CTX->totals.innocent_corpusfed     - s->control_totals.innocent_corpusfed) == 0 &&
+          abs(CTX->totals.spam_classified        - s->control_totals.spam_classified) == 0 &&
+          abs(CTX->totals.innocent_classified    - s->control_totals.innocent_classified) == 0)) {
 
-    result = PQexec(s->dbh, query);
+      buffer *buf;
+      buf = buffer_create (NULL);
+      if (buf == NULL) {
+        LOG (LOG_CRIT, ERR_MEM_ALLOC);
+        if (CTX->flags & DSF_MERGED)
+          memcpy(&CTX->totals, &user, sizeof(struct _ds_spam_totals));
+        return EUNKNOWN;
+      }
 
-    if ( !result || (PQresultStatus(result) != PGRES_COMMAND_OK && PQresultStatus(result) != PGRES_NONFATAL_ERROR) )
-    {
-      _pgsql_drv_query_error (PQresultErrorMessage(result), query);
-      if (result) PQclear(result);
-      if (CTX->flags & DSF_MERGED)
-        memcpy(&CTX->totals, &user, sizeof(struct _ds_spam_totals));
-      return EFAILURE;
+      snprintf (query, sizeof (query),
+                "UPDATE dspam_stats SET ");
+      buffer_copy (buf, query);
+
+      /* Do not update spam learned if no change is needed */
+      if (abs (CTX->totals.spam_learned - s->control_totals.spam_learned) != 0) {
+        snprintf (query, sizeof (query),
+                  "%sspam_learned=spam_learned%s%d",
+                  (update_any) ? "," : "",
+                  (CTX->totals.spam_learned > s->control_totals.spam_learned) ? "+" : "-",
+                  abs (CTX->totals.spam_learned - s->control_totals.spam_learned));
+        update_any = 1;
+        buffer_cat (buf, query);
+      }
+      /* Do not update innocent learned if no change is needed */
+      if (abs (CTX->totals.innocent_learned - s->control_totals.innocent_learned) != 0) {
+        snprintf (query, sizeof (query),
+                  "%sinnocent_learned=innocent_learned%s%d",
+                  (update_any) ? "," : "",
+                  (CTX->totals.innocent_learned > s->control_totals.innocent_learned) ? "+" : "-",
+                  abs (CTX->totals.innocent_learned - s->control_totals.innocent_learned));
+        update_any = 1;
+        buffer_cat (buf, query);
+      }
+      /* Do not update spam misclassified if no change is needed */
+      if (abs (CTX->totals.spam_misclassified - s->control_totals.spam_misclassified) != 0) {
+        snprintf (query, sizeof (query),
+                  "%sspam_misclassified=spam_misclassified%s%d",
+                  (update_any) ? "," : "",
+                  (CTX->totals.spam_misclassified > s->control_totals.spam_misclassified) ? "+" : "-",
+                  abs (CTX->totals.spam_misclassified - s->control_totals.spam_misclassified));
+        update_any = 1;
+        buffer_cat (buf, query);
+      }
+      /* Do not update innocent misclassified if no change is needed */
+      if (abs (CTX->totals.innocent_misclassified - s->control_totals.innocent_misclassified) != 0) {
+        snprintf (query, sizeof (query),
+                  "%sinnocent_misclassified=innocent_misclassified%s%d",
+                  (update_any) ? "," : "",
+                  (CTX->totals.innocent_misclassified > s->control_totals.innocent_misclassified) ? "+" : "-",
+                  abs (CTX->totals.innocent_misclassified - s->control_totals.innocent_misclassified));
+        update_any = 1;
+        buffer_cat (buf, query);
+      }
+      /* Do not update spam corpusfed if no change is needed */
+      if (abs (CTX->totals.spam_corpusfed - s->control_totals.spam_corpusfed) != 0) {
+        snprintf (query, sizeof (query),
+                  "%sspam_corpusfed=spam_corpusfed%s%d",
+                  (update_any) ? "," : "",
+                  (CTX->totals.spam_corpusfed > s->control_totals.spam_corpusfed) ? "+" : "-",
+                  abs (CTX->totals.spam_corpusfed - s->control_totals.spam_corpusfed));
+        update_any = 1;
+        buffer_cat (buf, query);
+      }
+      /* Do not update innocent corpusfed if no change is needed */
+      if (abs (CTX->totals.innocent_corpusfed - s->control_totals.innocent_corpusfed) != 0) {
+        snprintf (query, sizeof (query),
+                  "%sinnocent_corpusfed=innocent_corpusfed%s%d",
+                  (update_any) ? "," : "",
+                  (CTX->totals.innocent_corpusfed > s->control_totals.innocent_corpusfed) ? "+" : "-",
+                  abs (CTX->totals.innocent_corpusfed - s->control_totals.innocent_corpusfed));
+        update_any = 1;
+        buffer_cat (buf, query);
+      }
+      /* Do not update spam classified if no change is needed */
+      if (abs (CTX->totals.spam_classified - s->control_totals.spam_classified) != 0) {
+        snprintf (query, sizeof (query),
+                  "%sspam_classified=spam_classified%s%d",
+                  (update_any) ? "," : "",
+                  (CTX->totals.spam_classified > s->control_totals.spam_classified) ? "+" : "-",
+                  abs (CTX->totals.spam_classified - s->control_totals.spam_classified));
+        update_any = 1;
+        buffer_cat (buf, query);
+      }
+      /* Do not update innocent classified if no change is needed */
+      if (abs (CTX->totals.innocent_classified - s->control_totals.innocent_classified) != 0) {
+        snprintf (query, sizeof (query),
+                  "%sinnocent_classified=innocent_classified%s%d",
+                  (update_any) ? "," : "",
+                  (CTX->totals.innocent_classified > s->control_totals.innocent_classified) ? "+" : "-",
+                  abs (CTX->totals.innocent_classified - s->control_totals.innocent_classified));
+        buffer_cat (buf, query);
+      }
+
+      snprintf (query, sizeof (query),
+                " WHERE uid=%d",
+                (int) p->pw_uid);
+      buffer_cat (buf, query);
+
+      result = PQexec(s->dbh, buf->data);
+
+      if ( !result || (PQresultStatus(result) != PGRES_COMMAND_OK && PQresultStatus(result) != PGRES_NONFATAL_ERROR) )
+      {
+        _pgsql_drv_query_error (PQresultErrorMessage(result), buf->data);
+        if (result) PQclear(result);
+        if (CTX->flags & DSF_MERGED)
+          memcpy(&CTX->totals, &user, sizeof(struct _ds_spam_totals));
+        buffer_destroy (buf);
+        return EFAILURE;
+      }
+      buffer_destroy (buf);
     }
   }
 
@@ -776,7 +845,8 @@ _ds_setall_spamrecords (DSPAM_CTX * CTX, ds_diction_t diction)
             s->pg_token_type == 0 ? "numeric" : "bigint");
   buffer_cat (prepare, scratch);
 
-  if ((abs (control.spam_hits - s->control_sh)) != 0)
+  /* Do not update spam hits if no change is needed */
+  if (abs(control.spam_hits - s->control_sh) != 0)
   {
     if (control.spam_hits > s->control_sh)
     {
@@ -794,7 +864,8 @@ _ds_setall_spamrecords (DSPAM_CTX * CTX, ds_diction_t diction)
     buffer_cat (prepare, scratch);
   }
 
-  if ((abs (control.innocent_hits - s->control_ih)) != 0)
+  /* Do not update innocent hits if no change is needed */
+  if (abs(control.innocent_hits - s->control_ih) != 0)
   {
     if (control.innocent_hits > s->control_ih)
     {
@@ -1317,7 +1388,7 @@ _ds_get_signature (DSPAM_CTX * CTX, struct _ds_spam_signature *SIG,
   unsigned char *mem, *mem2;
   char query[256];
   PGresult *result;
-  int uid;
+  int uid = -1;
 
   if (s->dbh == NULL)
   {
@@ -1373,7 +1444,9 @@ _ds_get_signature (DSPAM_CTX * CTX, struct _ds_spam_signature *SIG,
     CTX->username = username;
     _ds_init_storage(CTX, (dbh_attached) ? dbh : NULL);
     s = (struct _pgsql_drv_storage *) CTX->storage;
-  } else {
+  }
+
+  if (uid == -1) {
     uid = (int) p->pw_uid;
   }
 
@@ -1618,7 +1691,7 @@ _ds_get_nextuser (DSPAM_CTX * CTX)
 #ifndef VIRTUAL_USERS
   struct passwd *p;
 #else
-  char *virtual_table, *virtual_uid, *virtual_username;
+  char *virtual_table, *virtual_username;
 #endif
   uid_t uid;
   char query[256];
@@ -1633,10 +1706,6 @@ _ds_get_nextuser (DSPAM_CTX * CTX)
   if ((virtual_table
     = _ds_read_attribute(CTX->config->attributes, "PgSQLVirtualTable"))==NULL)
   { virtual_table = "dspam_virtual_uids"; }
-
-  if ((virtual_uid
-    = _ds_read_attribute(CTX->config->attributes, "PgSQLVirtualUIDField"))==NULL)
-  { virtual_uid = "uid"; }
 
   if ((virtual_username = _ds_read_attribute(CTX->config->attributes,
     "PgSQLVirtualUsernameField")) ==NULL)
@@ -1659,7 +1728,9 @@ _ds_get_nextuser (DSPAM_CTX * CTX)
       if (result) PQclear(result);
       return NULL;
     }
-    if (result) PQclear(result);    result = NULL;
+    if (result)
+      PQclear(result);
+    result = NULL;
 
     /* Declare Cursor */
 #ifdef VIRTUAL_USERS
@@ -1789,8 +1860,7 @@ _ds_get_nexttoken (DSPAM_CTX * CTX)
       if (result) PQclear(result);
       result = PQexec(s->dbh, "END");
       if (result) PQclear(result);
-      free(st);
-      return NULL;
+      goto FAIL;
     }
     if (result) PQclear(result);
 
@@ -1809,8 +1879,7 @@ _ds_get_nexttoken (DSPAM_CTX * CTX)
       if (result) PQclear(result);
       result = PQexec(s->dbh, "END");
       if (result) PQclear(result);
-      free(st);
-      return NULL;
+      goto FAIL;
     }
     if (result) PQclear(result);
 
@@ -1826,8 +1895,7 @@ _ds_get_nexttoken (DSPAM_CTX * CTX)
     if (result) PQclear(result);
     if (s->iter_token) PQclear(s->iter_token);
     s->iter_token = NULL;
-    free(st);
-    return NULL;
+    goto FAIL;
   }
 
   if ( PQntuples(s->iter_token) < 1 ) {
@@ -1837,26 +1905,31 @@ _ds_get_nexttoken (DSPAM_CTX * CTX)
     if (result) PQclear(result);
     if (s->iter_token) PQclear(s->iter_token);
     s->iter_token = NULL;
-    free(st);
-    return NULL;
+    goto FAIL;
   }
 
   st->token = _pgsql_drv_token_read (s->pg_token_type, PQgetvalue( s->iter_token, 0, 0));
   st->spam_hits = strtoul (PQgetvalue( s->iter_token, 0, 1), NULL, 0);
   if (st->spam_hits == ULONG_MAX && errno == ERANGE) {
     LOGDEBUG("_ds_get_nexttoken: failed converting %s to st->spam_hits", PQgetvalue(s->iter_token,0,1));
-    free(st);
-    return NULL;
+    goto FAIL;
   }
   st->innocent_hits = strtoul (PQgetvalue( s->iter_token, 0, 2), NULL, 0);
   if (st->innocent_hits == ULONG_MAX && errno == ERANGE) {
     LOGDEBUG("_ds_get_nexttoken: failed converting %s to st->innocent_hits", PQgetvalue(s->iter_token,0,2));
-    free(st);
-    return NULL;
+    goto FAIL;
   }
   st->last_hit = (time_t) strtol ( PQgetvalue( s->iter_token, 0, 3), NULL, 0);
+  if (st->last_hit == ULONG_MAX && errno == ERANGE) {
+    LOGDEBUG("_ds_get_nexttoken: failed converting %s to st->last_hit", PQgetvalue(s->iter_token,0,3));
+    goto FAIL;
+  }
 
   return st;
+
+FAIL:
+  free(st);
+  return NULL;
 }
 
 /*
@@ -1924,8 +1997,7 @@ _ds_get_nextsignature (DSPAM_CTX * CTX)
       if (result) PQclear(result);
       result = PQexec(s->dbh, "END");
       if (result) PQclear(result);
-      free(st);
-      return NULL;
+      goto FAIL;
     }
     if (result) PQclear(result);
 
@@ -1944,8 +2016,7 @@ _ds_get_nextsignature (DSPAM_CTX * CTX)
       if (result) PQclear(result);
       result = PQexec(s->dbh, "END");
       if (result) PQclear(result);
-      free(st);
-      return NULL;
+      goto FAIL;
     }
     if (result) PQclear(result);
 
@@ -1961,8 +2032,7 @@ _ds_get_nextsignature (DSPAM_CTX * CTX)
     if (result) PQclear(result);
     if (s->iter_sig) PQclear(s->iter_sig);
     s->iter_sig = NULL;
-    free(st);
-    return NULL;
+    goto FAIL;
   }
 
   if ( PQntuples(s->iter_sig) < 1 ) {
@@ -1972,13 +2042,13 @@ _ds_get_nextsignature (DSPAM_CTX * CTX)
     if (result) PQclear(result);
     if (s->iter_sig) PQclear(s->iter_sig);
     s->iter_sig = NULL;
-    free(st);
-    return NULL;
+    goto FAIL;
   }
 
   if ( PQgetlength(s->iter_sig, 0, 0) == 0 ) {
-    free(st);
-    return NULL;
+    if (s->iter_sig) PQclear(s->iter_sig);
+    s->iter_sig = NULL;
+    goto FAIL;
   }
 
   mem = PQunescapeBytea( (unsigned char *) PQgetvalue( s->iter_sig, 0, 0), &length );
@@ -1987,8 +2057,7 @@ _ds_get_nextsignature (DSPAM_CTX * CTX)
   if (st->data == NULL) {
     LOG(LOG_CRIT, ERR_MEM_ALLOC);
     PQFREEMEM(mem);
-    free(st);
-    return NULL;
+    goto FAIL;
   }
 
   memcpy(st->data, mem, length);
@@ -1998,18 +2067,20 @@ _ds_get_nextsignature (DSPAM_CTX * CTX)
   if (st->length == LONG_MAX && errno == ERANGE) {
     LOGDEBUG("_ds_get_nextsignature: failed converting %s to st->length", PQgetvalue(s->iter_sig,0,2));
     free(st->data);
-    free(st);
-    return NULL;
+    goto FAIL;
   }
   st->created_on = (time_t) strtol (PQgetvalue(s->iter_sig, 0, 3), NULL, 0);
   if (st->created_on == LONG_MAX && errno == ERANGE) {
     LOGDEBUG("_ds_get_nextsignature: failed converting %s to st->created_on", PQgetvalue(s->iter_sig,0,3));
     free(st->data);
-    free(st);
-    return NULL;
+    goto FAIL;
   }
 
   return st;
+
+FAIL:
+  free(st);
+  return NULL;
 }
 
 struct passwd *
@@ -2261,6 +2332,9 @@ _pgsql_drv_query_error (const char *error, const char *query)
 struct passwd *
 _pgsql_drv_setpwnam (DSPAM_CTX * CTX, const char *name)
 {
+  if (name == NULL)
+    return NULL;
+
   char query[256];
   char *virtual_table, *virtual_uid, *virtual_username;
   struct _pgsql_drv_storage *s = (struct _pgsql_drv_storage *) CTX->storage;
@@ -2651,6 +2725,15 @@ int _ds_pref_set (
     m2 = PQescapeBytea((unsigned char *) value, strlen(value), &length);
   }
 
+  if (m1 == NULL || m2 == NULL)
+  {
+    LOG (LOG_CRIT, ERR_MEM_ALLOC);
+    dspam_destroy(CTX);
+    PQFREEMEM(m1);
+    PQFREEMEM(m2);
+    return EFAILURE;
+  }
+
   snprintf(query, sizeof(query), "DELETE FROM dspam_preferences"
     " WHERE uid=%d AND preference='%s'", (int) uid, m1);
 
@@ -2722,7 +2805,7 @@ int _ds_pref_del (
       LOGDEBUG ("_ds_pref_del: unable to _pgsql_drv_getpwnam(%s)",
               username);
       dspam_destroy(CTX);
-      return EUNKNOWN;
+      return EFAILURE;
     } else {
       uid = (int) p->pw_uid;
     }
@@ -2734,6 +2817,14 @@ int _ds_pref_del (
     m1 = PQescapeByteaConn(s->dbh, (unsigned char *) preference, strlen(preference), &length);
   } else {
     m1 = PQescapeBytea((unsigned char *) preference, strlen(preference), &length);
+  }
+
+  if (m1 == NULL)
+  {
+    LOG (LOG_CRIT, ERR_MEM_ALLOC);
+    dspam_destroy(CTX);
+    PQFREEMEM(m1);
+    return EFAILURE;
   }
 
   snprintf(query, sizeof(query), "DELETE FROM dspam_preferences"
